requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
Flask-SocketIO>=5.3.6
WTForms>=3.1.1
Werkzeug>=3.0.1
SQLAlchemy>=2.0.23
python-dotenv>=1.0.0
eventlet>=0.33.3
openpyxl>=3.1.2
email-validator>=2.1.0
markupsafe>=2.1.3